Passing msgs in blks so sweep, halo no dead lock
In the previous design, all producers would pass the entire message
before moving on to the consumer part code, this could fully fill up the
routing device and cause dead lock.
This commit adjust the location of the fine grain loop, so that a node
would switch between producer and consumer with partial message moved.