Auto-complete task switching trace
When tracing qthread, it happened that I missed some PCs that would
switch a task on. With this commit, I let the script try inference which
task is occupying the kthread based on a later yield event.
Some jump events could be inserted by trace2json.py to make it a bit
easier for json2svg.py to render a proper chart.